libsnap-confine-private: fix comparison of signed with unsigned

+++ cmd/libsnap-confine-private/utils.c
@@ -84,7 +84,8 @@ static int parse_bool(const char *text, bool * value, bool default_value)
 		*value = default_value;
 		return 0;
 	}
-	for (int i = 0; i < sizeof sc_bool_names / sizeof *sc_bool_names; ++i) {
+	for (size_t i = 0; i < sizeof sc_bool_names / sizeof *sc_bool_names;
+	     ++i) {
 		if (strcmp(text, sc_bool_names[i].text) == 0) {
 			*value = sc_bool_names[i].value;
 			return 0;
